fix: remove preview import cycle
