+++ bd_to_avp/modules/preview_range.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class PreviewRange:
+    start_seconds: float
+    duration_seconds: float
+    source_duration_seconds: float
